requirements.txt
accelerate==1.10.0
blobfile==3.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
diskcache==5.6.3
faiss-cpu==1.12.0
filelock==3.19.1
fsspec==2025.7.0
hf-xet==1.1.8
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.16
lxml==6.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.9
pycryptodomex==3.23.0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
triton==3.4.0
typing_extensions==4.14.1
urllib3==2.5.0
wheel==0.45.1
